requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
fastapi==0.117.1
h11==0.16.0
idna==3.10
paho-mqtt==2.1.0
pydantic==2.11.9
pydantic_core==2.33.2
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
